Initial commit
Contains initial implementation of a MailBox class along with other
inner classes for Folder and Message.
This is not ideal but I don't want to lose snipets using the lower level
APIs.

A better implementation of this might be subclassing the builtin
mailbox.Mailbox for use with an IMAP backend.