Refactor search and methods that take a message set
Previously search was returning a list of UIDs after decoding and
splitting a bytes value. But this list would later on be merged back
together using commas as the separator for use in other operations.
Now we just replace the spaces with commas and pass the UIDs around as
a bytes value.